# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  template_internal_http_datatransfers_requests.AdminCreateUserRequest:
    properties:
      first_name:
        maxLength: 100
        type: string
      first_name_en:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      last_name_en:
        maxLength: 100
        type: string
      register:
        maxLength: 20
        minLength: 6
        type: string
      role_id:
        enum:
        - 2
        - 3
        - 4
        type: integer
    required:
    - register
    type: object
...
  template_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Gerege Template Platform V3.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Gerege Template Platform V3.0 API
  version: "1.0"
paths:
...
  /v1/admin/users:
    post:
      consumes:
      - application/json
      description: Иргэнийг регистрийн дугаараар урьдчилан бүртгэж, role ононо. Private
        горимд зөвхөн ингэж бүртгэсэн иргэн Gerege SSO-оор нэвтэрнэ. 'users.manage'
        эрх шаардана; admin/superadmin role-ыг зөвхөн super admin ононо.
      parameters:
      - description: Register + role
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User pre-registered
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Insufficient privilege for the role
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "409":
          description: Register already exists
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Хэрэглэгч урьдчилан бүртгэх (private платформ)
      tags:
      - admin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"